this should be a uint32 instead of int16 for num triangles
change cast

+++ stl.cpp
@@ -391,7 +391,7 @@ int stl::read_ascii()
         }
     } while (tok != nullptr);
 
-    m_num_triangles = static_cast<int16_t>(m_vectors.size() / 9);
+    m_num_triangles = static_cast<uint32_t>(m_vectors.size() / 9);
     return result;
 }
